chore: aggregate curve, cleanup deps
