+++ package.json
@@ -27,9 +27,7 @@
   "dependencies": {
     "@upstash/redis": "^1.35.3",
     "axios": "^1.6.2",
-    "csv-parse": "^6.1.0",
     "dotenv": "^16.3.1",
-    "lodash": "^4.17.21",
     "node-cache": "^5.1.2",
     "p-limit": "^4.0.0",
     "viem": "^2.37.6",
@@ -38,7 +36,6 @@
   "devDependencies": {
     "@biomejs/biome": "^2.2.2",
     "@types/jest": "^29.5.11",
-    "@types/lodash": "^4.17.20",
     "@types/node": "^20.10.5",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
